Add methods to initialize config directory and setup all credentials into the dialog
The method which creates config directory was added. This method initializes config directory if it doesn't exist.
Also echo mode to hide writing a password was added.
During the object is creating the config loading method is called. This method loads all important settings from the configuration file. After that, all settings will be set into the dialog window.