Add methods to process any operations with queries
At first, when you want to make a query, you have to use the add_query method to add the query into the query list. Then you have to use the execute_query method to execute the query. If you want to reject a query, you may use the reject_query method to remove all or only one query from the query list.

The rest changes in the other methods were made due to the added methods.